+++ packages/backend-function/src/factory.ts
@@ -464,7 +464,7 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
   };
 
   private resolveRuntime = () => {
-    const runtimeDefault = 20;
+    const runtimeDefault = 22;
 
     // if runtime is not set, default to the oldest LTS
     if (!this.props.runtime) {
@@ -651,23 +651,15 @@ class AmplifyFunction
   ) {
     super(scope, id, outputStorageStrategy);
 
-    const runtime = nodeVersionMap[props.runtime];
-
     const require = createRequire(import.meta.url);
 
-    const shims =
-      runtime === Runtime.NODEJS_16_X
-        ? []
-        : [require.resolve('./lambda-shims/cjs_shim')];
+    const shims = [require.resolve('./lambda-shims/cjs_shim')];
 
     const ssmResolverFile =
-      runtime === Runtime.NODEJS_16_X
-        ? require.resolve('./lambda-shims/resolve_ssm_params_sdk_v2') // use aws cdk v2 in node 16
-        : require.resolve('./lambda-shims/resolve_ssm_params');
+      require.resolve('./lambda-shims/resolve_ssm_params');
 
-    const invokeSsmResolverFile = require.resolve(
-      './lambda-shims/invoke_ssm_shim',
-    );
+    const invokeSsmResolverFile =
+      require.resolve('./lambda-shims/invoke_ssm_shim');
 
     /**
      * This code concatenates the contents of the ssm resolver and invoker into a single line that can be used as the esbuild banner content
@@ -827,13 +819,13 @@ const isWholeNumberBetweenInclusive = (
   max: number,
 ) => min <= test && test <= max && test % 1 === 0;
 
-export type NodeVersion = 16 | 18 | 20 | 22;
+export type NodeVersion = 18 | 20 | 22 | 24;
 
 const nodeVersionMap: Record<NodeVersion, Runtime> = {
-  16: Runtime.NODEJS_16_X,
   18: Runtime.NODEJS_18_X,
   20: Runtime.NODEJS_20_X,
   22: Runtime.NODEJS_22_X,
+  24: Runtime.NODEJS_24_X,
 };
 
 export type FunctionArchitecture = 'x86_64' | 'arm64';