+++ packages/backend-function/src/provided_function_factory.ts
@@ -35,9 +35,7 @@ export type ProvidedFunctionProps = {
 /**
  * Adapts provided CDK function as Amplify function.
  */
-export class ProvidedFunctionFactory
-  implements ConstructFactory<AmplifyFunction>
-{
+export class ProvidedFunctionFactory implements ConstructFactory<AmplifyFunction> {
   private generator: ConstructContainerEntryGenerator;
 
   /**