+++ packages/deployed-backend-client/src/stack-name-resolvers/app_name_and_branch_main_stack_name_resolver.ts
@@ -13,9 +13,7 @@ export type AppNameAndBranchBackendIdentifier = {
 /**
  * Resolves stack names given an Amplify app name and branch
  */
-export class AppNameAndBranchMainStackNameResolver
-  implements MainStackNameResolver
-{
+export class AppNameAndBranchMainStackNameResolver implements MainStackNameResolver {
   /**
    * Initialize with an Amplify client and app info
    */