+++ packages/deployed-backend-client/src/stack_metadata_output_retrieval_strategy.ts
@@ -21,9 +21,7 @@ import {
 /**
  * Gets Amplify backend outputs from stack metadata and outputs
  */
-export class StackMetadataBackendOutputRetrievalStrategy
-  implements BackendOutputRetrievalStrategy
-{
+export class StackMetadataBackendOutputRetrievalStrategy implements BackendOutputRetrievalStrategy {
   /**
    * Instantiate with a CloudFormationClient and a StackNameResolver
    */