+++ packages/integration-tests/src/test-in-memory/data_storage_auth_with_triggers.test.ts
@@ -55,7 +55,6 @@ void it('data storage auth with triggers', async () => {
     'defaultNodeFunctionlambda5C194062',
     'echoFunclambdaE17DCA46',
     'handler2lambda1B9C7EFF',
-    'node16Functionlambda97ECC775',
     'onUploadlambdaA252C959',
     'onDeletelambda96BB6F15',
   ]);