+++ packages/platform-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/platform-core
 
+## 1.11.0
+
+### Minor Changes
+
+- 67e8773: Add standalone deployment type for deploying Gen2 backends without Amplify Hosting
+
+### Patch Changes
+
+- Updated dependencies [0ee9189]
+- Updated dependencies [67e8773]
+  - @aws-amplify/plugin-types@1.12.0
+
 ## 1.10.4
 
 ### Patch Changes